# swagger.yaml
openapi: 3.0.0
info:
  title: Group Management API
  description: API for managing groups and participants
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /group:
    post:
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                description: ID of the created group
    get:
      summary: Get all groups
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    description:
                      type: string
  /group/{id}:
    get:
      summary: Get group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
                  participants:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        wish:
                          type: string
    put:
      summary: Update group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      summary: Delete group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /group/{id}/participant:
    post:
      summary: Add participant to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                wish:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: integer
                description: ID of the added participant
  /group/{groupId}/participant/{participantId}:
    delete:
      summary: Delete participant from a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /group/{id}/toss:
    post:
      summary: Perform a toss within a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    wish:
                      type: string
                    recipient:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        wish:
                          type: string
  /group/{groupId}/participant/{participantId}/recipient:
    get:
      summary: Get recipient for a participant in a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: participantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  wish:
                    type: string